+++ resources/leiningen/new/gossamer/src/util.clj
@@ -4,10 +4,3 @@
 
 
 (log/deflogger metrics "METRICS")
-
-
-(defn handle-uncaught-exception
-  "Callback function for handling uncaught exceptions."
-  [^Thread thread ^Throwable ex]
-  (log/error (Throwable->map ex) ex (format "Uncaught exception on thread ID: %d, thread name: %s - (%s) %s"
-                                      (.getId thread) (.getName thread) (class ex) (.getMessage ex))))