+++ CHANGELOG.md
@@ -13,6 +13,7 @@ None
   - bract-context.edn
   - bract-context.dev.edn
 - Drop namespaces `app.command` and `app.keydef`
+- Drop utility fn `handle-uncaught-exception`
 
 
 ## 0.1.4 / 2019-January-06